+++ src/main/java/ru/gb/pugacheva/homework3springboot/homework3springboot/services/ProductService.java
@@ -24,7 +24,7 @@ public class ProductService {
         productRepository.add(product);
     }
 
-    public Product findById (Long id){
+    public Product findById (int id){
         return productRepository.findById(id);
     }
 }
